myapp/requirements.txt
strawberry-graphql==0.155.4
requests==2.28.2
flask==2.2.2
redis==4.5.1
